Support for multiple resume variants
Stylesheet now takes $variant param to define which version of the
resume is desired.

So far, only the sequence of bullet points under a job description can
be customized, by using the new <job-items>/<job-item-list>/<job-item>
elements (the names could probably be improved here) to specify which
bullet points belong to which resume variant and in what order.

Includes a first cut at mobile-first and general variants for Cantina.